requirements.txt
aiohttp==3.7.3
beautifulsoup4==4.9.3
discord.py==1.6.0
PyPDF2==1.26.0
validate-docbr==1.8.2
